move asymmetric signature/verification interfaces
